Moved fixtures to confttest file to remove clutter from unittest modules
