+++ tests/profile/serializers/test_indicator_data_serializer.py
@@ -1,42 +1,7 @@
 import pytest
 
-from tests.profile.factories import ProfileFactory, ProfileIndicatorFactory
-from tests.datasets.factories import GeographyFactory, IndicatorDataFactory, MetaDataFactory
-
 from wazimap_ng.profile.serializers.indicator_data_serializer import get_indicator_data, get_profile_data
 
-@pytest.fixture
-def profile():
-    return ProfileFactory()
-
-@pytest.fixture
-def geography():
-    return GeographyFactory()
-
-@pytest.fixture
-def profile_indicators(profile):
-    pi1 = ProfileIndicatorFactory(profile=profile, label="PI1")
-    pi2 = ProfileIndicatorFactory(profile=profile, label="PI2")
-    return [
-        pi1, pi2
-    ]
-
-@pytest.fixture
-def indicator_data(geography, profile_indicators):
-    idata = []
-
-    for pi in profile_indicators:
-        indicator = pi.indicator
-        idatum = IndicatorDataFactory(geography=geography, indicator=indicator)
-        idata.append(idatum)
-
-
-    return idata
-
-@pytest.fixture
-def metadata(indicator_data):
-    dataset = indicator_data[0].indicator.dataset
-    return MetaDataFactory(source="A source", url="http://example.com", description="A description", dataset=dataset)
 
 @pytest.mark.django_db
 @pytest.mark.usefixtures("indicator_data")
